move some functions to optimization_utils
